requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.7
bcrypt==4.0.1
loguru==0.7.2
pytest==7.4.4
httpx==0.26.0
python-dotenv==1.0.0 